refactor: rename query-filter.ts to query-filter-utils.ts
