+++ apps/frontend/src/providers/analytics/analytics-types.ts
@@ -1,7 +1,7 @@
 import type { Labrinth } from '@modrinth/api-client'
 import type { LocationQueryValueRaw } from 'vue-router'
 
-import type { ProjectStatusFilterValue } from '~/components/analytics-dashboard/query-builder/query-filter'
+import type { ProjectStatusFilterValue } from '~/components/analytics-dashboard/query-builder/query-filter-utils'
 
 export type AnalyticsQueryFilterCategory =
 	| 'project'
